refactor into separate crate
