+++ Cargo.toml
@@ -64,3 +64,4 @@ edgee-sdk = "1.2.0"
 edgee-path = "0.1.1"
 
 edgee-server = { version = "0.6.0", path = "crates/server" }
+wasmtime-runtime = { path = "crates/wasmtime-runtime"}
