+++ crates/cli/src/commands/serve.rs
@@ -1,7 +1,7 @@
 setup_command! {}
 
 pub async fn run(_opts: Options) {
-    edgee_server::init();
+    edgee_server::init().unwrap();
 
     tokio::select! {
         Err(err) = edgee_server::monitor::start() => tracing::error!(?err, "Monitor failed"),
