+++ crates/server/src/proxy/compute/data_collection/mod.rs
@@ -13,15 +13,13 @@ use http::{header, HeaderMap};
 use json_comments::StripComments;
 use regex::Regex;
 use tracing::{info, warn, Instrument};
+use wasmtime_runtime::components::{self};
+use wasmtime_runtime::payload::{self, EventData, EventType, Payload};
 
-use super::html::Document;
-use crate::config;
+use crate::{config, get};
+use crate::proxy::compute::html::Document;
 use crate::proxy::context::incoming::RequestHandle;
 use crate::tools::edgee_cookie;
-use payload::{EventData, EventType, Payload};
-
-pub mod components;
-pub mod payload;
 
 #[tracing::instrument(name = "data_collection", skip(document, request, response))]
 pub async fn process_from_html(
@@ -111,7 +109,8 @@ pub async fn process_from_html(
     // send the payload to the data collection components
     tokio::spawn(
         async move {
-            if let Err(err) = components::send_data_collection(&events).await {
+            let config = config::get();
+            if let Err(err) = components::send_data_collection(get(), &events, &config.components, &config.log.debug_component).await {
                 warn!(?err, "failed to send data collection payload");
             }
         }
@@ -207,7 +206,8 @@ pub async fn process_from_json(
     // send the payload to the data collection components
     tokio::spawn(
         async move {
-            if let Err(err) = components::send_data_collection(&events).await {
+            let config = config::get();
+            if let Err(err) = components::send_data_collection(get(), &events,&config.components, &config.log.debug_component).await {
                 warn!(?err, "failed to send data collection payload");
             }
         }
