+++ crates/wasmtime-runtime/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "wasmtime-runtime"
+version.workspace = true
+authors.workspace = true
+license.workspace = true
+keywords.workspace = true
+repository.workspace = true
+homepage.workspace = true
+edition.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+chrono = { workspace = true, features = ["serde"] }
+http.workspace = true
+json_pretty.workspace = true
+reqwest.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+tokio = { workspace = true}
+tracing.workspace = true
+uuid = { workspace = true, features = ["v4", "serde"] }
+wasmtime.workspace = true
+wasmtime-wasi.workspace = true
