+++ crates/wasmtime-runtime/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod components;
+pub mod payload;
+
+
+wasmtime::component::bindgen!({
+    world: "data-collection",
+    path: "wit/protocols.wit",
+    async: true,
+});
