+++ crates/wasmtime-runtime/wit/protocols.wit
